--- oap-data-source/arrow/src/main/scala/com/intel/oap/spark/sql/execution/datasources/v2/arrow/ArrowDataSourceV2.scala
@@ -1,41 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.intel.oap.spark.sql.execution.datasources.v2.arrow
-
-import com.intel.oap.spark.sql.execution.datasources.arrow.ArrowFileFormat
-
-import org.apache.spark.sql.connector.catalog.Table
-import org.apache.spark.sql.execution.datasources.FileFormat
-import org.apache.spark.sql.execution.datasources.v2.FileDataSourceV2
-import org.apache.spark.sql.util.CaseInsensitiveStringMap
-
-class ArrowDataSourceV2 extends FileDataSourceV2 {
-
-  private val format = classOf[ArrowFileFormat]
-
-  override def fallbackFileFormat: Class[_ <: FileFormat] = {
-    format
-  }
-
-  override def getTable(options: CaseInsensitiveStringMap): Table = {
-    val paths = getPaths(options)
-    val tableName = getTableName(paths)
-    ArrowTable(tableName, sparkSession, options, paths, None, fallbackFileFormat)
-  }
-
-  override def shortName(): String = "arrow"
-}
